Update README.md
A restaurant web application that allows customers to order food by scanning a QR code placed on their table. Once an order is placed, the request is sent in realtime to the restaurant managers dashboard for efficient processing. The system streamlines the ordering process, reduces wait times, and improves restaurant management.